flask_app/requirements.txt
flask==3.0.0
gunicorn==21.2.0
google-cloud-bigquery==3.25.0
pandas==2.1.0
pyarrow==14.0.1
db-dtypes==1.2.0